Condense edit playlist components
