+++ src/components/EditBox/EditBox.js
@@ -2,7 +2,6 @@ import React, { useContext, useRef } from 'react';
 import usePlaylist from '../../hooks/usePlaylist';
 import useSpotify from '../../hooks/useSpotify';
 import { PositionContext } from '../../context/PositionContext';
-import EditBoxTracks from '../EditBoxTracks/EditBoxTracks';
 
 import './EditBox.css';
 
@@ -12,7 +11,7 @@ const EditBox = () => {
   const { state } = useContext(PositionContext);
   const { playListPosition } = state;
   
-  const { editBoxIsOpen, editListPlayLists, editListTracks } = usePlaylist();
+  const { editBoxIsOpen, editListPlayLists, editListTracks, removeTrack } = usePlaylist();
   const { openPlayLists, updatePlayList } = useSpotify();
     
 
@@ -27,6 +26,35 @@ const EditBox = () => {
     updatePlayList(playListId, newName, uris);
   }
 
+  const renderEditTracks = () => {
+    if (Array.isArray(editListTracks) && editListTracks.length > 0) {
+      return (
+        <div className="TrackList">
+          {
+            editListTracks.map((track, index) => {
+              return (
+                <div key={`${track.uri}:${index}`} className="Track">
+                  <div className="Track-information" id="track">
+                    <h3>{track.name}</h3>
+                    <p>{track.artistName} | {track.albumName}</p>
+                  </div>
+                  <button className="Track-action"
+                    onClick={() => removeTrack(index, true)}>-</button>
+                </div>
+              );
+            })
+          }
+        </div>
+      );
+    } else {
+      return (
+        <div className="TrackList">
+          <br /><p>Add some tracks!</p>
+        </div>
+      );
+    }
+  }
+
   if (editBoxIsOpen === true) {
       return (
         <div className="Editlist">
@@ -38,8 +66,7 @@ const EditBox = () => {
           <button className="Editlist-save" onClick={() => handleUpdateClick(editListPlayLists[playListPosition].id)}>
           <b>UPDATE ON SPOTIFY</b></button>
           <div className="Track-counter">Number of tracks: {editListTracks.length}</div>
-          <EditBoxTracks />
-          {/* <Pagination /> */}
+          { renderEditTracks() }
         </div>
       );
   }
