+++ src/components/ListOfPlayLists/ListOfPlayLists.css
@@ -39,6 +39,34 @@
   overflow-y: scroll;
 }
 
+.Playlist-list .List-item {
+  display: flex;
+  justify-content: space-between;
+  width: 88%;
+  margin: 2% 0 0 4%;
+  padding: 2%;
+  border: 2px solid #2943e6;
+}
+
+.Playlist-list .List-item p {
+  max-width: 75%;
+}
+
+.Title {
+  color: #2ec56a;
+}
+
+.Playlist-list .List-item:hover {
+  border-color: #fff;
+}
+
+.Count {
+  color: rgba(251, 251, 251, 0.62);
+}
+
+.Playlist-list .List-item:hover .Count {
+  color: rgb(251, 251, 251);
+}
 
 .Edit-btn {
   color: rgba(251, 251, 251, 0.62);
